Extracted the subnavitems to its own file
