+++ src/views/GovernanceReviewTeam/sideNavItems.js
@@ -1,0 +1,33 @@
+const subNavItems = [
+  {
+    page: `intake-request`,
+    text: 'general:intake'
+  },
+  {
+    page: `business-case`,
+    text: 'general:businessCase'
+  },
+  {
+    page: `decision`,
+    text: 'decision.title'
+  },
+  {
+    page: `lcid`,
+    text: 'lifecycleID.title',
+    groupEnd: true
+  },
+  {
+    page: `actions`,
+    text: 'actions'
+  },
+  {
+    page: `notes`,
+    text: 'notes.heading'
+  },
+  {
+    page: `dates`,
+    text: 'dates.heading'
+  }
+];
+
+export default subNavItems;
